MAGECLOUD-1568: Create integration test for MAGETWO-72296
