+++ src/Test/Integration/Bootstrap.php
@@ -66,13 +66,6 @@ class Bootstrap
             $buildFile,
             $sandboxDir . '/build_options.ini'
         ));
-        /**
-         * Install without dev dependencies.
-         */
-        $this->execute(sprintf(
-            'composer install -n -d %s --no-dev --no-progress',
-            $sandboxDir
-        ));
     }
 
     /**
